Fix test performance comparison
We were running the `largeJavaMultiProject` with `max-workers=4`
and `maxParallelForks=4`, which lead to each project spawning 4
test workers (for only 100 tests) and thus blocking all other projects
from testing. Instead, we now use `maxParallelForks=1` for this scenario,
which is much more realistic for a build that has lots of projects with
only a few tests each.